pyproject.toml
[tool.poetry]
name = "GoldBot"
version = "0.1.0"
description = "RAG pipelines with Ploomber"
authors = ["Rajat Jain rjain032@ucr.edu"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ploomber = "^0.23.0"
ipykernel = "^6.25.2"
farm-haystack = {extras = ["inference","faiss"], version = "^1.20.1"}
python-dotenv = "^1.0.0"
ipywidgets = "^8.1.1"
chainlit = "^0.7.0"
pandas = "2.0.3"
yfinance = "^0.2.30"
matplotlib = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chainlit = "chainlit run src/app/app.py"
